Implement logic to handle catch-22 wrt. continuous read mode
Idea is to issue a select command and wait for a response the usual way.
Then we issue a time-out read to see if any other UID's arrive and if so
wait for 'S' response end-of-list identifier and carry on as usual.

Also, don't bother reading the EEPROM every time we connect to the device
cos we'll probably be better off trying to figure out device state as we
go or just read specific registers as required.

It ought to be noted that some operating modes can be changed without
modifying the EEPROM registers so it'll be more reliable that way anyway.